+++ frontend/src/css/department/Department_info.module.css
