+++ frontend/src/css/department/group/Group_info_list.module.css
