+++ frontend/src/css/personnel/detail/Personnel_detail.module.css
